reworked project file structure
