+++ webapp/src/http_handlers/sessions/mod.rs
@@ -2,3 +2,6 @@
 pub mod session;
 // Anonymous user sessions tracking
 pub mod sessions_tracker;
+
+pub use session::Session;
+pub use sessions_tracker::SessionsTracker;
