+++ webapp/src/http_handlers/submit/structs/submit_input.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct SubmitInput
+{
+    pub lang: String,
+    pub options: String,
+    pub code: String
+}
